changing plan
change the usage of API.

The ICMP api was too constraining, it also blocked if it come from the same host so I passed to the w2s api insted to make the raw packets.

added a new function to receive echo or repply echo ( might change it once I use multiple host wouldn't want to use twice the same data ).