Good point to stop for day 2
at this point the software can

send ICMP with custome data using windows socket

receive ICMP and read the information inside of the ICMP packet.

it can generate checksum for a payload ( not connected yet).

we have the information on how to make the TCP flags coded and the maximum size.  by math

a way to do padding so we can get the information

we will need to add in information of the readme (use for note at the moment ) to how we will keep active connections (JSON might be a idea?) and we could fake a header name " port " to be able to have multiple connection ? we will see let make it work with one conection for now and be happy !