starting TCP logic
everything was moved into an object so we can manage multiple connection easier in the future.

right now every object can do one TCP connection using a MAP with "ip" and "last_flag" to help with the information

some formating was added for the sequance number

a function to make the sequance number a integer again was made

starting logic with packet receive (such as a Syn, synack,ack command. received.